Move loadApiDisplayInputs out of bl.cpp
Extract ApiDisplayInputs assembly into display_session so bl.cpp no longer
owns preference/wakeup/battery field packing for /api/display.

Move vBatt into globals for the pre-WiFi voltage snapshot shared by the
session and device status stamps. Behavior unchanged.

Verified: pio run -e trmnl
